brooklyn-example-simple-nosql-cluster: add org.apache package prefix
