+++ examples/simple-nosql-cluster/src/main/java/org/apache/brooklyn/demo/SimpleRedisCluster.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package brooklyn.demo;
+package org.apache.brooklyn.demo;
 
 import brooklyn.entity.basic.ApplicationBuilder;
 import brooklyn.entity.nosql.redis.RedisCluster;
