+++ examples/simple-nosql-cluster/src/main/java/org/apache/brooklyn/demo/StormSampleApp.java
@@ -16,43 +16,55 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package brooklyn.demo;
+package org.apache.brooklyn.demo;
 
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import brooklyn.catalog.Catalog;
 import brooklyn.entity.basic.AbstractApplication;
 import brooklyn.entity.basic.Entities;
 import brooklyn.entity.basic.StartableApplication;
-import brooklyn.entity.nosql.cassandra.CassandraDatacenter;
+import brooklyn.entity.messaging.storm.StormDeployment;
 import brooklyn.entity.proxying.EntitySpec;
 import brooklyn.launcher.BrooklynLauncher;
 import brooklyn.util.CommandLineUtil;
 
 import com.google.common.collect.Lists;
 
-public class SimpleCassandraCluster extends AbstractApplication {
+/**
+ * Sample showing a storm analyticscluster
+ **/
+@Catalog(name="Storm Sample App",
+description="Creates a Storm analytics cluster",
+    iconUrl="classpath://brooklyn/demo/glossy-3d-blue-web-icon.png")
+public class StormSampleApp extends AbstractApplication implements StartableApplication {
 
-    private static final String DEFAULT_LOCATION = "localhost";
+    public static final Logger LOG = LoggerFactory.getLogger(StormSampleApp.class);
+
+    public static final String DEFAULT_LOCATION = "named:gce-europe-west1";
 
     @Override
     public void initApp() {
-        addChild(EntitySpec.create(CassandraDatacenter.class)
-                .configure(CassandraDatacenter.INITIAL_SIZE, 1)
-                .configure(CassandraDatacenter.CLUSTER_NAME, "Brooklyn"));
+        addChild(EntitySpec.create(StormDeployment.class)
+            .configure(StormDeployment.SUPERVISORS_COUNT, 2)
+            .configure(StormDeployment.ZOOKEEPERS_COUNT, 1));
     }
-    
+
     public static void main(String[] argv) {
         List<String> args = Lists.newArrayList(argv);
         String port =  CommandLineUtil.getCommandLineOption(args, "--port", "8081+");
         String location = CommandLineUtil.getCommandLineOption(args, "--location", DEFAULT_LOCATION);
-        
+
         BrooklynLauncher launcher = BrooklynLauncher.newInstance()
-                 .application(EntitySpec.create(StartableApplication.class, SimpleCassandraCluster.class)
-                         .displayName("Cassandra"))
-                         .webconsolePort(port)
-                 .location(location)
-                 .start();
-             
+            .application(EntitySpec.create(StartableApplication.class, StormSampleApp.class)
+                .displayName("Storm App"))
+                .webconsolePort(port)
+                .location(location)
+                .start();
+
         Entities.dumpInfo(launcher.getApplications());
     }
 }
